requirements.txt
streamlit>=1.47.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0